split base/parented view provider classes into separate modules; split tests; remove duplicated ParentedBaseViewProvider tests for search
